compliance-agent-system/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
boto3>=1.34.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
colorama>=0.4.6
tqdm>=4.66.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
mcp>=0.1.0
fastmcp>=0.1.0
httpx>=0.24.0
uvicorn>=0.23.0
sse-starlette>=1.6.0